refactor: move make_callback to utils
